requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
dill==0.4.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.68.3
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
llvmlite==0.45.1
MarkupSafe==3.0.2
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.3
packaging==25.0
pillow==12.0.0
psutil==7.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.2.5
pyspnego==0.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qiskit==2.1.2
qiskit-ibm-runtime==0.41.1
requests==2.32.5
requests_ntlm==1.3.0
rustworkx==0.17.1
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
stevedore==5.5.0
sympy==1.14.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
